+++ wagtailmenus/models.py
@@ -180,7 +180,7 @@ class MainMenu(ClusterableModel):
         verbose_name_plural = _("main menu")
 
     @classmethod
-    def for_site(cls, site):
+    def get_for_site(cls, site):
         """
         Get a mainmenu instance for the site.
         """
@@ -217,6 +217,23 @@ class FlatMenu(ClusterableModel):
         unique_together = ("site", "handle")
         verbose_name = _("flat menu")
         verbose_name_plural = _("flat menus")
+
+    @classmethod
+    def get_for_site(cls, handle, site,
+                     fall_back_to_default_site_menus=False):
+        """
+        Get a FlatMenu instance with a matching `handle` for the `site`
+        provided - or for the 'default' site if not found.
+        """
+        menu = cls.objects.filter(handle__exact=handle, site=site).first()
+        if(
+            menu is None and fall_back_to_default_site_menus and
+            not site.is_default_site
+        ):
+            return cls.objects.filter(
+                handle__exact=handle, site__is_default_site=True
+            ).first()
+        return menu
 
     def __str__(self):
         return '%s (%s)' % (self.title, self.handle)