+++ CHANGELOG.md
@@ -1,13 +1,37 @@
 Changelog
 =========
 
-1.4.1 (XX.XX.XXXX) IN DEVELOPMENT
+1.5.0 (XX.XX.XXX) IN DEVELOPMENT
 ---------------------------------
+* Updated FlatMenu listing in CMS to only show site column and filters if menus
+  are defined for more than one site.
+* Added the `fall_back_to_default_site_menus` option to the `flat_menu` tag, to
+  allow flat menus defined for the default site to be used as fall-backs, in
+  cases where the 'current' site doesn't have its own menus set up with the
+  specified handle.
+* Apply `active` classes to menu items that link to custom URLs (if
+  `request.path` and `link_url` are exact matches).
 * Added a `handle` to `MenuItem` model to provide a string which can be 
   used to do specific matching of menu items in the template. (Tim Leguijt)
 
+
+1.4.1 (02.10.2016) 
+------------------
+
+* Updated FlatMenu listing in CMS to include a column for `site`, a filter for
+  `handle`, and a MenuItem count for each object. Also added default ordering,
+  and output the handle value in a <code></code> tag to make it stand out.
+* Made it easier to develop and debug wagtailmenus locally, by running it as a
+  Django project. See CONTRIBUTING.md for instructions. 
+* Added a `get_for_site` class method to the FlatMenu model, to be consistent 
+  with the MainMenu model, and renamed the `for_site` method on MainMenu to
+  `get_for_site` for consistency. `main_menu` and `flat_menu` tags now make use
+  of these.
+* Fixed an minor bug in the `prime_menu_items` method, where a `depth`
+  value was hard-coded, instead of utilising the `SECTION_ROOT_LEVEL` setting. 
+
 1.4.0 (22.09.2016)
----------------------------------
+------------------
 
 * Added a `has_submenu_items()` method to `MenuPage` model to compliment
   `modify_submenu_items()` in version 1.3. Allows for far better control and 
@@ -28,7 +52,7 @@ Changelog
 
 
 1.3.1 (09.08.2016)
----------------------------------
+------------------
 
 * Configured additional tox test environments for Wagtail>=1.6 with Django=1.9
   and 1.10.
@@ -39,7 +63,7 @@ Changelog
   where possible.
 
 1.3.0 (06.08.2016)
----------------------------------
+------------------
 
 * Added the ability for all menu tags to (attempt to) identify ancestor pages
   and section root page by using components from the request path when serving
@@ -51,7 +75,7 @@ Changelog
   
 
 1.2.3 (25.07.2016)
----------------------------------
+------------------
 
 * Added PyPi version and coveralls test coverage badges to README
 * Altered Travis CI test configuration to use tox, allowing for much better